move  to utils and use it to fix sidebar labels
